--- deployment/ccip/config/rmn.go
@@ -1,39 +1,0 @@
-package config
-
-import (
-	"math/big"
-
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/rmn_home"
-	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/rmn_remote"
-	"github.com/smartcontractkit/chainlink/v2/core/services/keystore/keys/p2pkey"
-)
-
-type RMNNopConfig struct {
-	NodeIndex           uint64
-	OffchainPublicKey   [32]byte
-	EVMOnChainPublicKey common.Address
-	PeerId              p2pkey.PeerID
-}
-
-func (c RMNNopConfig) ToRMNHomeNode() rmn_home.RMNHomeNode {
-	return rmn_home.RMNHomeNode{
-		PeerId:            c.PeerId,
-		OffchainPublicKey: c.OffchainPublicKey,
-	}
-}
-
-func (c RMNNopConfig) ToRMNRemoteSigner() rmn_remote.RMNRemoteSigner {
-	return rmn_remote.RMNRemoteSigner{
-		OnchainPublicKey: c.EVMOnChainPublicKey,
-		NodeIndex:        c.NodeIndex,
-	}
-}
-
-func (c RMNNopConfig) SetBit(bitmap *big.Int, value bool) {
-	if value {
-		bitmap.SetBit(bitmap, int(c.NodeIndex), 1)
-	} else {
-		bitmap.SetBit(bitmap, int(c.NodeIndex), 0)
-	}
-}
